requirements.txt
google-api-python-client
numpy
opencv-python
pytesseract
google_auth_oauthlib
dateutils
